Improve behavior of the sequential subscriber
When processing of a message fails in a batch, we can immediately nack
all following messages in the batch, since the stream will fail. This
reduces time when they are locked in the queue, while we know they won't
be processed.

Moreover, all previous messages that were processed in the same batch
can be emitted, as we know they were successful.
This second part is similar to what `evalMap` does, but the current
approach keeps the chunk structure (`evalMap` unchunks everything),
which can lead to better performances.