Expose the `QueuePusher` API
Similarly to the `QueuePuller` API defined in #6, the publication API is
modeled around a sink and a `QueuePusher`. The pusher is the lowest
level API of publication, allowing existing code bases to choose how the
messages are published. The sink provided in the `QueuePubliser` allows
for integrating into existing fs2 code easily.

This approach brings regularity to the API and symmetry between
publication and subscription.