Make access do deserialized payload effectful
Deserializing the content might fail if the payload is malformed. The
idea is to make access to it effectful (with memoization) so that
failure occurs when the payload is accessed. The logic behind it is that
this is a business logic error to handle, not something that can be
handled at the library level.

This PR also introduces access to the `rawPayload` to handle malformed
body.